+++ src/audiovideocontroller/AudioVideoControllerState.ts
@@ -106,8 +106,11 @@ export default class AudioVideoControllerState {
   // itself correspond to transceivers; 0 in this list corresponds to an inactive tranceiver.
   videoSubscriptions: number[] | null = null;
 
-  // Total number of videos streams that can be subscribed to.
-  videoSubscriptionLimit: number = 0;
+  // The video subscription limit is set by the backend and is subject to change in future.
+  // This value is set in the `JoinAndReceiveIndexTask` when we process the `SdkJoinAckFrame` 
+  // and is used in the `ReceiveVideoStreamIndexTask` to limit the total number of streams 
+  // that we include in the `videosToReceive`.
+  videoSubscriptionLimit: number = 25;
 
   videosPaused: VideoStreamIdSet | null = null;