+++ CHANGELOG.md
@@ -7,15 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Ability to choose remote video sources in `AllHighestVideoBandwidthPolicy`
+- Ability to choose remote video sources in `AllHighestVideoBandwidthPolicy`. see [guide](https://aws.github.io/amazon-chime-sdk-js/modules/videolayout.html#downlink-policy)
 
 ### Removed
 
 ### Fixed
 
 ### Changed
-- Enforced a video receive limit incase the number of videos shared in the meeting
-are greater than the limit.
+- Enforced a video receive limit incase the number of videos shared in the meeting are greater than the limit. The current limit is 25, which can change in future.
 
 ## [2.24.0] - 2021-12-07
 ### Added